fixed style

+++ source/Ped.cpp
@@ -375,14 +375,14 @@ namespace GTA
 	}
 	int Ped::CreateGroup(int Unused)
 	{
-		return Native::Function::Call<int>(Native::Hash::CREATE_GROUP,this->Handle,(int)Unused);
+		return Native::Function::Call<int>(Native::Hash::CREATE_GROUP, this->Handle, Unused);
 	}
 	void Ped::SetPedAsGroupLeader(int GroupID)
 	{
-		Native::Function::Call(Native::Hash::SET_PED_AS_GROUP_LEADER, this, (int)GroupID);
+		Native::Function::Call(Native::Hash::SET_PED_AS_GROUP_LEADER, this, GroupID);
 	}
 	void Ped::SetPedAsGroupMember(int GroupID) {
-		Native::Function::Call(Native::Hash::SET_PED_AS_GROUP_MEMBER,this,(int)GroupID);
+		Native::Function::Call(Native::Hash::SET_PED_AS_GROUP_MEMBER, this, GroupID);
 	}
 	void Ped::RemovePedFromGroup() {
 		Native::Function::Call(Native::Hash::REMOVE_PED_FROM_GROUP, this);
